isfile check added to assemble_cat_results

+++ wikiDumpParser/processorResults.py
@@ -26,11 +26,12 @@ class ProcessorResults:
             path = os.path.join(self.project.results_path, key[:-3])
             f = 'cats.csv.7z'
             f = self.unpack(path, f)
-            data = pd.read_csv(os.path.join(path, f), header=None, delimiter='\t', na_filter=False)
-            # self.cats = self.cats.append(data)
-            results = os.path.join(self.project.data_path, 'cats_all.csv')
-            data.to_csv(results, sep='\t', index=False, header=False, mode='a')
-            os.remove(os.path.join(path, f))
+            if os.path.isfile(f):
+                data = pd.read_csv(os.path.join(path, f), header=None, delimiter='\t', na_filter=False)
+                # self.cats = self.cats.append(data)
+                results = os.path.join(self.project.data_path, 'cats_all.csv')
+                data.to_csv(results, sep='\t', index=False, header=False, mode='a')
+                os.remove(os.path.join(path, f))
 
     def remove_duplicate_authors(self):
         authors_file = os.path.join(self.project.results_path, 'author_info.csv')
